+++ packages/dash-core-components/src/components/DatePickerRange.react.js
@@ -1,9 +1,9 @@
 import {DateRangePicker} from 'react-dates';
-import moment from 'moment';
 import PropTypes from 'prop-types';
 import R from 'ramda';
 import React, {Component} from 'react';
-import './css/react-dates@12.3.0.css';
+
+import convertToMoment from '../utils/convertToMoment';
 
 /**
  * DatePickerRange is a tailor made component designed for selecting
@@ -30,22 +30,15 @@ export default class DatePickerRange extends Component {
          * - user modifiable attributes
          * - moment converted attributes
          */
-        const newState = {};
-        const momentProps = [
+
+        const newState = convertToMoment(newProps, [
             'start_date',
             'end_date',
             'initial_visible_month',
             'max_date_allowed',
             'min_date_allowed',
-        ];
-        momentProps.forEach(prop => {
-            if (R.type(newProps[prop]) !== 'Undefined') {
-                newState[prop] = moment(newProps[prop]);
-            }
-            if (prop === 'max_date_allowed' && R.has(prop, newState)) {
-                newState[prop].add(1, 'days');
-            }
-        });
+        ]);
+
         this.setState(newState);
     }