+++ packages/dash-core-components/src/components/DatePickerSingle.react.js
@@ -4,6 +4,8 @@ import PropTypes from 'prop-types';
 import R from 'ramda';
 import React, {Component} from 'react';
 
+import convertToMoment from '../utils/convertToMoment';
+
 /**
  * DatePickerSingle is a tailor made component designed for selecting
  * a single day off of a calendar.
@@ -15,6 +17,7 @@ import React, {Component} from 'react';
  * This component is based off of Airbnb's react-dates react component
  * which can be found here: https://github.com/airbnb/react-dates
  */
+
 export default class DatePickerSingle extends Component {
     constructor() {
         super();
@@ -29,21 +32,14 @@ export default class DatePickerSingle extends Component {
          * - user modifiable attributes
          * - moment converted attributes
          */
-        const newState = {};
-        const momentProps = [
+
+        const newState = convertToMoment(newProps, [
             'date',
             'initial_visible_month',
             'max_date_allowed',
             'min_date_allowed',
-        ];
-        momentProps.forEach(prop => {
-            if (R.type(newProps[prop]) !== 'Undefined') {
-                newState[prop] = moment(newProps[prop]);
-            }
-            if (prop === 'max_date_allowed' && R.has(prop, newState)) {
-                newState[prop].add(1, 'days');
-            }
-        });
+        ]);
+
         this.setState(newState);
     }